+++ src/screens/index.ts
@@ -1,17 +1,17 @@
 export { WelcomeScreen } from './WelcomeScreen';
 export { ConnectScreen } from './ConnectScreen';
+export { DeviceAuthScreen } from './DeviceAuthScreen';
 export { ConfigureScreen } from './ConfigureScreen';
 export { NetworkScanScreen } from './NetworkScanScreen';
 export { CredentialsScreen } from './CredentialsScreen';
 export { ConnectingScreen } from './ConnectingScreen';
 export { SuccessScreen } from './SuccessScreen';
-export { ManageScreen } from './ManageScreen';
 
 export type { WelcomeScreenProps } from './WelcomeScreen';
 export type { ConnectScreenProps } from './ConnectScreen';
+export type { DeviceAuthScreenProps } from './DeviceAuthScreen';
 export type { ConfigureScreenProps } from './ConfigureScreen';
 export type { NetworkScanScreenProps } from './NetworkScanScreen';
 export type { CredentialsScreenProps } from './CredentialsScreen';
 export type { ConnectingScreenProps } from './ConnectingScreen';
 export type { SuccessScreenProps } from './SuccessScreen';
-export type { ManageScreenProps } from './ManageScreen';